Check return code for presence of mongod

+++ tools/clean-deploy.sh
@@ -50,8 +50,11 @@ sudo apt-get -q -yy purge cloud-installer
 
 # just make sure juju-mongodb died. LP#1306315
 MONGOD=$(pgrep mongod)
-if [ ${MONGOD} -gt 0 ]; then
-  sudo kill -9 ${MONGOD}
+RET=$?
+if [ ${RET} -eq 0 ]; then
+  if [ ${MONGOD} -gt 0 ]; then
+    sudo kill -9 ${MONGOD}
+  fi
 fi
 
 sudo rm -rf ~/.juju ~/.cloud-install || true
